requirements.txt
numpy
pandas
scipy
hnswlib
psycopg2-binary
streamlit
langchain
langchain-community
openpyxl
chromadb
pypdf
python-docx
google-generativeai
langchain_google_genai
langchain_chroma